5.04

+++ src/main/java/com/winter/yubin/board/notice/NoticeController.java
@@ -13,11 +13,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.ModelAndView;
 
 import com.winter.yubin.board.BoardDTO;
 import com.winter.yubin.file.FileDTO;
+import com.winter.yubin.member.MemberDTO;
 import com.winter.yubin.pager.Pager;
 
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -29,7 +32,7 @@ public class NoticeController {
 	@Autowired
 	private NoticeService noticeService;
 	
-	@Value("${yubin.board.notice}")
+	@Value("${app.board.notice}")
 	private String name;
 	
 	@ModelAttribute("name")
@@ -74,7 +77,7 @@ public class NoticeController {
 	
 	@GetMapping("create")
 	public String create()throws Exception{
-		return "board/create";
+		return "board/form-";
 	}
 
 	@PostMapping("create")
@@ -97,16 +100,34 @@ public class NoticeController {
 	}
 	
 	@PostMapping("update")
-	public String update(NoticeDTO noticeDTO, @RequestParam("attach") MultipartFile [] attach)throws Exception{
+	public ModelAndView update(NoticeDTO noticeDTO, @RequestParam("attach") MultipartFile [] attach, Model model)throws Exception{
 		int result = noticeService.update(noticeDTO, attach);
 		
-		return "redirect:./list";
+		
+		ModelAndView mv = new ModelAndView();
+		
+		mv.setViewName("redirect:./list");
+		
+		mv.addObject("dto", noticeDTO);
+		
+		return mv;
 	}
 	
 	@PostMapping("delete")
-	public String delete(NoticeDTO noticeDTO)throws Exception{
-		int result = noticeService.delete(noticeDTO);
-		return "redirect:./list";
+	public String delete(NoticeDTO noticeDTO, HttpSession session, Model model)throws Exception{
+		MemberDTO memberDTO = (MemberDTO)session.getAttribute("member");
+		BoardDTO boardDTO = noticeService.detail(noticeDTO);
+		if(boardDTO.getBoardWriter().equals(memberDTO.getUsername())) {
+			int result = noticeService.delete(noticeDTO);
+			return "redirect:./list";
+			
+		}else {
+			model.addAttribute("result", "작성자가 아님");
+			model.addAttribute("url", "./list");
+			return "commons/result";
+		}
+		
+		
 	}
